GP2-2218 WTE tab rearrangement
